Mbgljs textsize (#49)
* Handle MBGLJS perpendicular offset

Seems to fail with QGIS labelled line (default settings) exported to Geostyler > MBGLJS

* Cast text-size to float

I am getting errors in MBGLJS output, because `text-size` is quoted. This fixes this via `float()`. The bug needs to be confirmed, as does the fix, or another method for converting to a number should be used.