Add support for CIMMarkerPlacementAtExtremities
This is for oriented markers at the end of lines (typically arrow heads). We already supported the cases where one marker was either at the end or at the beginning of a line (CIMMarkerPlacementAtRatioPositions with position 0 or 1). In ArcGIS, there is additionally the case where one symbolizer specifies one marker at both ends of the line (CIMMarkerPlacementAtExtremities), so we need to created two symbolizers for this one.