Fix whereClause (labels)
The previous test for checking the validity of a property name (isalpha()) was too strict and would fail for instance if an underscore was present. It's replaced with a regex.
Additionnally the convertWhereClause was removing all parentheses in the expression including those in the property value. This is now limited to the opening and closing parentheses at the beginining and at the end of the expression.

Additionnally there are cases in a lyrx where the maximum or minimum scale is NaN so we test that the value is numeric before adding it to the rule.